+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- CHANGELOG NEWEST -->
 - ADD non-premium console.log on the [LokiJS RxStorage](https://rxdb.info/rx-storage-lokijs.html#disabling-the-non-premium-console-log)
 - FIX randomly failing replication test
+- ADD dedicated [errors page](https://rxdb.info/errors.html)
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->